Update README.md (#15)
* Update README.md

added badges

* added license badge and fixed newline

fixed newline semantic error

* Update README.md

added "downloads" badge. NOTE: This is automatic badge. it will be updated automatically based on total downloads for EVERY RELEASE. Please @mention me if you wanted to to set it to specific version release.

* Update README.md

* Update README.md

fixed issue badge (this includes pull request issues too), so I replace it to specific pull request issue

* Update README.md

decided to make it total issues including pull request issues. Although, the link is to to the /issues and not /pull.